backend/requirements.txt
Django>=5.0,<7
djangorestframework>=3.15
psycopg2-binary>=2.9
celery[redis]>=5.3
redis>=5.0
django-cors-headers>=4.3
dj-database-url>=2.1
pytest>=8.0
pytest-django>=4.8
gunicorn>=22.0
